requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0
librosa>=0.10.0
soundfile>=0.12.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
xgboost>=2.0.0
shap>=0.42.0
tqdm>=4.65.0
# Note: parselmouth should be installed separately due to dependency issues
# Install with: pip install parselmouth --no-deps
# Or: pip install parselmouth --no-build-isolation

